add base class for storage tests
